Enable ImDebugger everywhere, on top of the regular UI.
